SwitchMultiPos Enhancement for Swtiches.h
The current version of the SwitchMultiPos has an issue for non-shorting multi-position rotary switches when one of the positions isn't connected to an Arduino pin.  While using PIN_NC to default to a position that's not hooked up tp a pin the switch's DCSBios output bounces between the default position and the next selected pin.  In the middle of a rotation step none of the pins are connected to ground.  By adding debounce logic to the multi-position switch , the time when the switch is between positions is ignored.